+++ src/context/index.js
@@ -1,0 +1,6 @@
+import { useWishlist } from "context/wishlist-context";
+import { useProduct } from "context/product-context";
+import { useCart } from "context/cart-context";
+import { useAuth } from "./auth-context";
+
+export { useAuth, useCart, useProduct, useWishlist };
